requirements.txt
alabaster==0.7.12
altgraph==0.17
appdirs==1.4.3
asgiref==3.3.4
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2019.11.28
cffi==1.15.0
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cryptography==36.0.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.0
Django==3.2.9
django-allauth==0.47.0
django-bootstrap4==21.1
django-cleanup==5.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-markdownx==3.0.1
django-oauth-toolkit==1.6.3
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
filelock==3.0.12
Flask==2.0.0
Flask-Cors==3.0.10
future==0.18.2
idna==2.10
imagesize==1.3.0
importlib-metadata==4.10.1
install==1.3.5
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==0.14.1
jwcrypto==1.0
kiwisolver==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.0
mecab==0.996.3
networkx==2.6.3
oauthlib==3.1.0
packaging==21.3
pefile==2019.4.18
Pillow==8.4.0
pipenv==2020.11.15
pycparser==2.21
Pygments==2.10.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyJWT==1.7.1
pymongo==3.12.3
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python3-openid==3.2.0
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rstcheck==3.3.1
six==1.14.0
snooty-lextudio==1.11.5
snowballstemmer==2.2.0
sorl-thumbnail==12.7.0
soupsieve==2.3.1
Sphinx==4.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
toml==0.10.2
tweepy==3.10.0
typing_extensions==4.0.1
urllib3==1.26.4
virtualenv==20.0.10
virtualenv-clone==0.5.3
watchdog==1.0.2
Werkzeug==2.0.0
wrapt==1.13.3
zipp==3.7.0
